Move from harmony_utils to serving_response
